Refactor : adopt interface type changed
