+++ src/app/recruit/[id]/edit/page.tsx
@@ -16,18 +16,11 @@ import useSWR from 'swr'
 import useAxiosWithAuth from '@/api/config'
 import { useRouter } from 'next/navigation'
 import RowRadioButtonsGroupStatus from './panel/radioGroupStatus'
-import { ITag, statusEnum } from '@/types/IPostDetail'
+import { IFormInterview, ITag, statusEnum } from '@/types/IPostDetail'
 
 export interface IRoleData {
   role: string | null
   member: number
-}
-
-export interface IFormInterview {
-  question: string
-  type: string
-  optionList?: string[]
-  ratioList?: { max: string; valueOfMin: string; valueOfMax: string }
 }
 
 const CreateTeam = ({ params }: { params: { recruit_id: string } }) => {
